requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.3.1
    # via
    #   keras
    #   tensorboard
    #   tensorflow
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
astunparse==1.6.3
    # via tensorflow
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
azure-core==1.35.0
    # via azure-storage-blob
azure-storage-blob==12.26.0
    # via -r requirements.in
beautifulsoup4==4.13.4
    # via yfinance
blinker==1.9.0
    # via flask
boto3==1.40.3
    # via -r requirements.in
botocore==1.40.3
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   curl-cffi
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   curl-cffi
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
cryptography==45.0.6
    # via azure-storage-blob
curl-cffi==0.12.0
    # via yfinance
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==6.0.1
    # via -r requirements.in
flatbuffers==25.2.10
    # via tensorflow
frozendict==2.4.6
    # via yfinance
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gast==0.6.0
    # via tensorflow
google-api-core==2.25.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.2.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-pasta==0.2.0
    # via tensorflow
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
grpcio==1.74.0
    # via
    #   tensorboard
    #   tensorflow
gunicorn==23.0.0
    # via -r requirements.in
h5py==3.14.0
    # via
    #   keras
    #   tensorflow
idna==3.10
    # via
    #   requests
    #   yarl
isodate==0.7.2
    # via azure-storage-blob
itsdangerous==2.2.0
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.6
    # via flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.1
    # via scikit-learn
keras==3.11.1
    # via tensorflow
libclang==18.1.1
    # via tensorflow
markdown==3.8.2
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.3
    # via
    #   keras
    #   tensorflow
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
multitasking==0.0.12
    # via yfinance
namex==0.1.0
    # via keras
numpy==2.1.3
    # via
    #   -r requirements.in
    #   h5py
    #   keras
    #   ml-dtypes
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
    #   yfinance
opt-einsum==3.4.0
    # via tensorflow
optree==0.17.0
    # via keras
packaging==25.0
    # via
    #   gunicorn
    #   keras
    #   tensorboard
    #   tensorflow
pandas==2.3.1
    # via
    #   -r requirements.in
    #   yfinance
peewee==3.18.2
    # via yfinance
platformdirs==4.3.8
    # via yfinance
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via google-api-core
protobuf==5.29.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   tensorboard
    #   tensorflow
    #   yfinance
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via
    #   pandas
    #   yfinance
requests==2.32.4
    # via
    #   azure-core
    #   google-api-core
    #   google-cloud-storage
    #   tensorflow
    #   yfinance
rich==13.9.4
    # via
    #   -r requirements.in
    #   keras
rsa==4.9.1
    # via google-auth
s3transfer==0.13.1
    # via boto3
scikit-learn==1.7.1
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
six==1.17.0
    # via
    #   astunparse
    #   azure-core
    #   google-pasta
    #   python-dateutil
    #   tensorboard
    #   tensorflow
soupsieve==2.7
    # via beautifulsoup4
structlog==25.4.0
    # via -r requirements.in
tensorboard==2.19.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.19.0
    # via -r requirements.in
tensorflow-io-gcs-filesystem==0.37.1
    # via tensorflow
termcolor==3.1.0
    # via tensorflow
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   azure-core
    #   azure-storage-blob
    #   beautifulsoup4
    #   multidict
    #   optree
    #   rich
    #   structlog
    #   tensorflow
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   botocore
    #   requests
websockets==15.0.1
    # via yfinance
werkzeug==3.1.3
    # via
    #   -r requirements.in
    #   flask
    #   flask-cors
    #   tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.2
    # via tensorflow
yarl==1.20.1
    # via aiohttp
yfinance==0.2.65
    # via -r requirements.in
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==1.10.15

# The following packages are considered to be unsafe in a requirements file:
# setuptools
